refactor(authflow): split login by authentication mechanism
Following zitadel's login layout (one handler file per mechanism:
password_handler, external_provider_handler, ...), split the 584-line login.go
into focused files: login.go (entry + connector dispatch), password.go (the
password-credential mechanism), callback.go (the redirect-return mechanism for
OAuth2/SAML), and finalize.go (the post-authentication step shared by all
mechanisms). No behaviour change.
